+++ core/app_client.go
@@ -16,6 +16,7 @@ package core
 
 import (
 	"application/core/model"
+	"application/driven/calendar"
 	"time"
 
 	"github.com/google/uuid"
@@ -35,13 +36,13 @@ func (a appClient) GetSurvey(id string, orgID string, appID string) (*model.Surv
 }
 
 // GetSurvey returns surveys matching the provided query
-func (a appClient) GetSurveys(orgID string, appID string, creatorID *string, surveyIDs []string, surveyTypes []string, limit *int, offset *int) ([]model.Survey, error) {
-	return a.app.shared.getSurveys(orgID, appID, creatorID, surveyIDs, surveyTypes, limit, offset)
+func (a appClient) GetSurveys(orgID string, appID string, creatorID *string, surveyIDs []string, surveyTypes []string, calendarEventID string, limit *int, offset *int) ([]model.Survey, error) {
+	return a.app.shared.getSurveys(orgID, appID, creatorID, surveyIDs, surveyTypes, calendarEventID, limit, offset)
 }
 
 // CreateSurvey creates a new survey
-func (a appClient) CreateSurvey(survey model.Survey) (*model.Survey, error) {
-	return a.app.shared.createSurvey(survey)
+func (a appClient) CreateSurvey(survey model.Survey, externalIDs map[string]string) (*model.Survey, error) {
+	return a.app.shared.createSurvey(survey, externalIDs)
 }
 
 // UpdateSurvey updates the provided survey
@@ -60,16 +61,113 @@ func (a appClient) GetSurveyResponse(id string, orgID string, appID string, user
 	return a.app.storage.GetSurveyResponse(id, orgID, appID, userID)
 }
 
-// GetSurveyResponses returns the survey responses matching the provided filters
-func (a appClient) GetSurveyResponses(orgID string, appID string, userID string, surveyIDs []string, surveyTypes []string, startDate *time.Time, endDate *time.Time, limit *int, offset *int) ([]model.SurveyResponse, error) {
+// GetUserSurveyResponses returns the survey responses matching the provided filters for a specific user
+func (a appClient) GetUserSurveyResponses(orgID string, appID string, userID string, surveyIDs []string, surveyTypes []string, startDate *time.Time, endDate *time.Time, limit *int, offset *int) ([]model.SurveyResponse, error) {
 	return a.app.storage.GetSurveyResponses(&orgID, &appID, &userID, surveyIDs, surveyTypes, startDate, endDate, limit, offset)
 }
 
+// GetAllSurveyResponses returns the survey responses matching the provided filters
+func (a appClient) GetAllSurveyResponses(orgID string, appID string, userID string, surveyID string, startDate *time.Time, endDate *time.Time, limit *int, offset *int, externalIDs map[string]string) ([]model.SurveyResponse, error) {
+	var allResponses []model.SurveyResponse
+	var err error
+
+	survey, err := a.app.shared.getSurvey(surveyID, orgID, appID)
+	if err != nil {
+		return nil, err
+	}
+
+	// Check if survey is sensitive
+	if survey.Sensitive {
+		return nil, errors.Newf("Survey is sensitive and responses are not available")
+	}
+
+	// If no calendar event is associated then user should not have access to respones
+	if len(survey.CalendarEventID) == 0 {
+		return nil, errors.Newf("Survey responses are not available. No calendar event associated")
+	}
+
+	// Get external ID
+	envConfig, err := a.app.GetEnvConfigs()
+	if err != nil {
+		return nil, errors.WrapErrorAction(logutils.ActionGet, model.TypeConfig, logutils.StringArgs(model.ConfigTypeEnv), err)
+	}
+	externalID := externalIDs[envConfig.ExternalID]
+	user := calendar.User{AccountID: userID, ExternalID: externalID}
+
+	// Check if user is admin of calendar event
+	eventUsers, err := a.app.calendar.GetEventUsers(survey.OrgID, survey.AppID, survey.CalendarEventID, []calendar.User{user}, nil, "admin", nil)
+	if err != nil {
+		return nil, err
+	}
+	for _, eventUser := range eventUsers {
+		if (eventUser.User.ExternalID == externalID || eventUser.User.AccountID == survey.CreatorID) && eventUser.Role == "admin" {
+			// Get responses
+			allResponses, err = a.app.storage.GetSurveyResponses(&orgID, &appID, nil, []string{surveyID}, nil, startDate, endDate, limit, offset)
+			if err != nil {
+				return nil, err
+			}
+
+			// If survey is anonymous strip userIDs
+			if survey.Anonymous {
+				for i := range allResponses {
+					allResponses[i].UserID = ""
+				}
+			}
+
+			return allResponses, nil
+		}
+	}
+
+	return nil, errors.Newf("account is not admin of calendar event")
+}
+
 // CreateSurveyResponse creates a new survey response
-func (a appClient) CreateSurveyResponse(surveyResponse model.SurveyResponse) (*model.SurveyResponse, error) {
+func (a appClient) CreateSurveyResponse(surveyResponse model.SurveyResponse, externalIDs map[string]string) (*model.SurveyResponse, error) {
 	surveyResponse.ID = uuid.NewString()
 	surveyResponse.DateCreated = time.Now().UTC()
 	surveyResponse.DateUpdated = nil
+
+	// Get survey from storage
+	survey, err := a.app.storage.GetSurvey(surveyResponse.Survey.ID, surveyResponse.OrgID, surveyResponse.AppID)
+	if err != nil {
+		return nil, err
+	}
+	// Populate survey with data from client request
+	survey.Data = surveyResponse.Survey.Data
+	survey.SurveyStats = surveyResponse.Survey.SurveyStats
+	survey.ResultJSON = surveyResponse.Survey.ResultJSON
+	surveyResponse.Survey = *survey
+
+	if len(survey.CalendarEventID) > 0 {
+		// check if user attended calendar event
+
+		// Get external ID
+		envConfig, err := a.app.GetEnvConfigs()
+		if err != nil {
+			return nil, errors.WrapErrorAction(logutils.ActionGet, model.TypeConfig, logutils.StringArgs(model.ConfigTypeEnv), err)
+		}
+		externalID := externalIDs[envConfig.ExternalID]
+
+		user := calendar.User{AccountID: surveyResponse.UserID, ExternalID: externalID}
+		attended := true
+		registered := true
+		// This API ignores the filters so the attendance must be manually checked
+		eventUsers, err := a.app.calendar.GetEventUsers(surveyResponse.OrgID, surveyResponse.AppID, survey.CalendarEventID, []calendar.User{user}, &registered, "", &attended)
+
+		if err != nil {
+			return nil, err
+		}
+		for _, eventUser := range eventUsers {
+			if (eventUser.User.AccountID == surveyResponse.UserID || eventUser.User.ExternalID == externalID) && eventUser.Attended {
+				return a.app.storage.CreateSurveyResponse(surveyResponse)
+			}
+		}
+		return nil, errors.Newf("account has not attended calendar event")
+		// if len(eventUsers) == 0 { // user has not attended
+		// 	return nil, errors.Newf("account has not attended calendar event")
+		// }
+	}
+
 	return a.app.storage.CreateSurveyResponse(surveyResponse)
 }