+++ core/application.go
@@ -54,6 +54,7 @@ type Application struct {
 
 	storage       interfaces.Storage
 	notifications interfaces.Notifications
+	calendar      interfaces.Calendar
 }
 
 // Start starts the core part of the application
@@ -77,8 +78,8 @@ func (a *Application) GetEnvConfigs() (*model.EnvConfigData, error) {
 }
 
 // NewApplication creates new Application
-func NewApplication(version string, build string, storage interfaces.Storage, notifications interfaces.Notifications, logger *logs.Logger) *Application {
-	application := Application{version: version, build: build, storage: storage, notifications: notifications, logger: logger}
+func NewApplication(version string, build string, storage interfaces.Storage, notifications interfaces.Notifications, calendar interfaces.Calendar, logger *logs.Logger) *Application {
+	application := Application{version: version, build: build, storage: storage, notifications: notifications, calendar: calendar, logger: logger}
 
 	//add the drivers ports/interfaces
 	application.Default = newAppDefault(&application)